Api/requirements.txt
django-rest-framework==0.1.0
djangorestframework==3.12.4
Django==3.2.5
djangorestframework-jwt==1.11.0
django-cors-headers==3.10.0
mysqlclient==2.0.1
